Add player input filter and re-prompt. Add welcome
Welcome alert with instructions added for when page content isn't loaded.
Added player input filter that screens out any invalid response.
The +a in the input line prevents the prompt returning null and causing
an error if the player clicks cancel, instead asking them for a valid
choice just like with any other invalid input. The filter looks for this extra
a, as arock, apaper, and ascissors.