Create basic functionality with buttons.
Computer selection code modified to take the choice from an array
containing all the options.
Player input tied to event listeners on the 3 buttons added to the
page.
When the player clicks the game now makes a selection for the computer,
compares it to the player choice, and outputs the choices and round
result to the console.