Add lizard and spock. Swap to array for win find.
Lizard and spock buttons were added to index.html.
The method of storing the "weapons" and determining the winner was changed
to a method based on nested arrays. The player selection returns onne of the weapon
names as a string.  In the script there are consts declared for each weapon,
containing an array which contains first the name as a string, then the weapons that
beat it, also as strings. The computer picks one of these arrays at random, and if
the player choice string is found in the array the player is the winner.