Add end game overlay. Add more descriptive text.
div added which gains an "active" class when one side reaches five points
This div prevents further play, and contains a button which when clicked
resets the scores and removes the active class. The active class sets the
fixed position div to cover the full screen.

The descriptive text additions add to the win/lose text on each round, by
describing how the victor wins. These values were added to the "weapons"
arrays at the index above the winning weapon they corresond to. To facilitate
this a function was added which finds the corresponding array to the
players choice